+++ crates/miden-protocol/src/transaction/kernel/tx_event_id.rs
@@ -58,6 +58,8 @@ pub enum TransactionEventId {
 
     InputNoteIndexLookup = INPUT_NOTE_INDEX_LOOKUP_ID,
 
+    TxBeforeBlockWitnessLoad = TX_BEFORE_BLOCK_WITNESS_LOAD_ID,
+
     AuthRequest = AUTH_REQUEST_ID,
 
     PrologueStart = PROLOGUE_START_ID,
@@ -129,6 +131,7 @@ impl TransactionEventId {
             Self::NoteAfterAddAsset => &NOTE_AFTER_ADD_ASSET_NAME,
             Self::NoteBeforeAddAttachment => &NOTE_BEFORE_ADD_ATTACHMENT_NAME,
             Self::InputNoteIndexLookup => &INPUT_NOTE_INDEX_LOOKUP_NAME,
+            Self::TxBeforeBlockWitnessLoad => &TX_BEFORE_BLOCK_WITNESS_LOAD_NAME,
             Self::AuthRequest => &AUTH_REQUEST_NAME,
             Self::PrologueStart => &PROLOGUE_START_NAME,
             Self::PrologueEnd => &PROLOGUE_END_NAME,
@@ -222,6 +225,8 @@ impl TryFrom<EventId> for TransactionEventId {
             NOTE_BEFORE_ADD_ATTACHMENT_ID => Ok(TransactionEventId::NoteBeforeAddAttachment),
 
             INPUT_NOTE_INDEX_LOOKUP_ID => Ok(TransactionEventId::InputNoteIndexLookup),
+
+            TX_BEFORE_BLOCK_WITNESS_LOAD_ID => Ok(TransactionEventId::TxBeforeBlockWitnessLoad),
 
             AUTH_REQUEST_ID => Ok(TransactionEventId::AuthRequest),